Add 'scope' option to set from Config
for DefaultProviders there is currently no way to change the scope for the authentification.

For Github it is needed to have org:read scope to limit, so only members can join the server. But in the current solution it only works if member are public members of that group. 

I think to make it overrideable in the config makes it more configurable for each ones need. 

As im not sure about security stuff i want to have comments on this.